pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fair-test"
description = "A library to build and deploy FAIR metrics tests APIs that can be used by FAIR evaluation services supporting the FAIRMetrics specifications, such as FAIR enough and the FAIR evaluator."
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
authors = [
    { name = "Vincent Emonet", email = "vincent.emonet@gmail.com" },
]
keywords = [
    "FAIR",
    "test",
    "evaluation",
    "API",
]
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development",
    "Typing :: Typed",
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: AsyncIO",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Internet :: WWW/HTTP",
]
dynamic = ["version"]

dependencies = [
    "fastapi >=0.51.0",
    "pydantic",
    "python-dotenv",
    "requests >=2.24.0,<3.0.0",
    "rdflib >=6.1.1",
    "PyLD",
    "extruct",
    "PyYAML >=5.3.1,<7.0.0",
    "idutils",
]


[project.optional-dependencies]
test = [
    "pytest >=7.1.3,<8.0.0",
    "pytest-cov >=2.12.0,<4.0.0",
    "mypy ==0.971",
    "black == 22.8.0",
    "isort >=5.0.6,<6.0.0",
    "flake8 >=3.8.3,<6.0.0",
    "Flake8-pyproject>=1.1.0.post0",
]
doc = [
    "mkdocs >=1.1.2,<2.0.0",
    "mkdocs-material >=8.1.4,<9.0.0",
    "mkdocstrings[python]",
    "mdx-include >=1.4.1,<2.0.0",
    "mkdocs-markdownextradata-plugin >=0.1.7,<0.3.0",
]
dev = [
    "mypy ==0.971",
    "uvicorn[standard] >=0.12.0,<0.19.0",
    "pre-commit >=2.17.0,<3.0.0",
    "autoflake >=1.4.0,<2.0.0",
    "flake8 >=3.8.3,<6.0.0",
    "Flake8-pyproject>=1.1.0.post0",

    "types-requests",
]


[project.urls]
Homepage = "https://github.com/MaastrichtU-IDS/fair-test"
Documentation = "https://maastrichtu-ids.github.io/fair-test"
History = "https://github.com/MaastrichtU-IDS/fair-test/releases"
Tracker = "https://github.com/MaastrichtU-IDS/fair-test/issues"
Source = "https://github.com/MaastrichtU-IDS/fair-test"



[tool.hatch.version]
path = "fair_test/__init__.py"


[tool.isort]
line_length = 120
profile = "black"
skip = ["fair_test/__init__.py"]
# known_third_party = ["fair_test"]


[tool.black]
color = true
line-length = 120
target-version = ['py310']
# skip-string-normalization = true


[tool.flake8]
max-line-length = 120
select = ["C", "E", "F", "W", "B", "B9"]
ignore = ["E203", "E501", "W503"]
exclude = ["__init__.py"]
# max-complexity = 14


[tool.mypy]
strict = false
disallow_untyped_defs = false
follow_imports = "normal"
# ignore_missing_imports = true
pretty = true
show_column_numbers = true
warn_no_return = false
warn_unused_ignores = true




# Trying out env from hatch https://hatch.pypa.io
[tool.hatch.envs.default]
features = [
  "test",
  "doc",
  "dev",
]
# post-install-commands = [
#   "pre-commit install",
# ]

# hatch run dev
[tool.hatch.envs.default.scripts]
dev = "./scripts/dev.sh"
format = "./scripts/format.sh"
lint = "./scripts/lint.sh"
mint = """black fair_test example tests
isort fair_test example tests"""
docs = "./scripts/docs-serve.sh"
test = "./scripts/test.sh"


[tool.hatch.envs.test.scripts]
all = "./scripts/test.sh"
# cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=fair_test --cov=tests"
# no-cov = "cov --no-cov"


[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true
# omit = [
#   "fair_test/__init__.py",
# ]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]



[tool.pytest.ini_options]
filterwarnings = [
    "ignore:the imp module is deprecated in favour of importlib.*:DeprecationWarning:pyRdfa",
    "ignore:It looks like you're parsing an XML document using an HTML parser.*:bs4.XMLParsedAsHTMLWarning:bs4"
]
# addopts = [
#   "--strict-config",
#   "--strict-markers",
# ]
# xfail_strict = true
# junit_family = "xunit2"
